Remove fix for attempting to pass bash variables to CLI
`crab echo $PORT` would yield `crab echo ` if not set. This can be special cased to also replace `PORT` with the port number, but this causes issues with replacing parts of words (i.e. 'REPORT'). We can remove this, as it's possible to wrap the variable in single quotes to achieve the desired outcome (`crab echo '$PORT')